+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 8. Move sidebar-content and related URL constants to `common-private`.
 9. Rename package to `@grailbio/lib`, publish to github and npm.
 10. filterResults uses default values if they aren't provided.
+11. filterResults can now filter and sort deeply nested objects.
 
 ## 1.0.0 - 2019-06-17